requirements.txt
streamlit
streamlit==1.44.1
pinecone==6.0.2
google-genai==1.9.0
sentence_transformers
